backend/pyproject.toml
[project]
name = "text-2-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt<4.0.0",
    "boto3>=1.43.6",
    "datasets>=4.8.4",
    "faker>=40.15.0",
    "fastapi>=0.136.1",
    "google-generativeai>=0.8.6",
    "langchain-aws>=1.4.6",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.2.1",
    "langfuse>=4.6.1",
    "mysql-connector-python>=9.6.0",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "sqlglot>=30.7.0",
    "torch>=2.11.0",
    "uvicorn>=0.46.0",
]
